// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  blurDataUrl?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "articles".
 */
export interface Article {
  id: number;
  title: string;
  slug?: string | null;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  contentSummary?: string | null;
  readTimeInMins?: number | null;
  coverImage: number | Media;
  author: number | ArticleAuthor;
  status?: ('Draft' | 'Published') | null;
  published_at?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "article-authors".
 */
export interface ArticleAuthor {
  id: number;
  name: string;
  role: 'Staff Writer' | 'Guest Writer' | 'Editor' | 'Contributor';
  avatar: number | Media;
  updatedAt: string;
  createdAt: string;
}